requirements.txt
streamlit
pillow
python-dotenv
openai
pymupdf
pandas